:sparkles: :: editor picks API 추가
