+++ core/data/src/main/java/com/skogkatt/data/datasource/article/ArticleDataSource.kt
@@ -2,9 +2,12 @@ package com.skogkatt.data.datasource.article
 
 import com.skogkatt.network.model.article.NetworkArticleResponse
 import com.skogkatt.network.model.article.NetworkContentResponse
+import com.skogkatt.network.model.editorspicks.NetworkEditorsPicksResponse
 
 internal interface ArticleDataSource {
     suspend fun getArticles(page: Int, query: String?, section: String?): NetworkArticleResponse
 
     suspend fun getArticleContent(id: String): NetworkContentResponse
+
+    suspend fun getEditorsPicks(): NetworkEditorsPicksResponse
 }
