+++ core/data/src/main/java/com/skogkatt/data/datasource/article/ArticleDataSourceImpl.kt
@@ -3,6 +3,7 @@ package com.skogkatt.data.datasource.article
 import com.skogkatt.network.api.retrofit.GuardianApi
 import com.skogkatt.network.model.article.NetworkArticleResponse
 import com.skogkatt.network.model.article.NetworkContentResponse
+import com.skogkatt.network.model.editorspicks.NetworkEditorsPicksResponse
 import javax.inject.Inject
 
 internal class ArticleDataSourceImpl @Inject constructor(
@@ -22,5 +23,9 @@ internal class ArticleDataSourceImpl @Inject constructor(
 
     override suspend fun getArticleContent(id: String): NetworkContentResponse {
         return guardianApi.getArticleContent(id)
+    }
+
+    override suspend fun getEditorsPicks(): NetworkEditorsPicksResponse {
+        return guardianApi.getEditorsPicks()
     }
 }
