+++ core/network/src/main/java/com/skogkatt/network/model/editorspicks/NetworkEditorsPicksResponse.kt
@@ -1,0 +1,7 @@
+package com.skogkatt.network.model.editorspicks
+
+import com.skogkatt.network.model.article.NetworkArticle
+
+data class NetworkEditorsPicksResponse(
+    val editorsPicks: List<NetworkArticle>,
+)
